parasitic values in ParasiticTab
git-svn-id: https://svn.java.net/svn/electric~svn/trunk@2505 b64ed6c1-75a8-e629-b1da-d0dae74a743a

+++ srcj/com/sun/electric/tool/extract/TransistorPBucket.java
@@ -68,8 +68,8 @@ public class TransistorPBucket implements ExtractedPBucket
         line.append(" " + TextUtils.formatDouble(ni.getAnchorCenterY()));
         if (type == 'n') line.append(" g=S_gnd");
         else line.append(" g=S_vdd"); // type = 'p'
-        line.append(" s=A_" + activeArea + ",P_" + activePerim);
-        line.append(" d=A_" + activeArea + ",P_" + activePerim);
+        line.append(" s=A_" + (int)activeArea + ",P_" + (int)activePerim);
+        line.append(" d=A_" + (int)activeArea + ",P_" + (int)activePerim);
         return line.toString();
     }
 }
